+++ tangy-base-widget.js
@@ -1,8 +1,6 @@
 import {html, PolymerElement} from '@polymer/polymer/polymer-element.js'
 import '@polymer/paper-card/paper-card.js'
 import '@polymer/paper-button/paper-button.js'
-import {Fab} from '@material/mwc-fab'
-import {Icon} from "@material/mwc-icon"
 
 const MODE_INFO = 'MODE_INFO'
 const MODE_EDIT = 'MODE_EDIT'
@@ -191,10 +189,10 @@ class TangyBaseWidget extends PolymerElement {
     return `
       <tangy-input 
         name="name" 
-        valid-if="input.value.match(/^[a-zA-Z].{1,}[a-zA-Z0-9\-_]$/)" 
+        valid-if="input.value.match(/^[a-zA-Z]{1,}[a-zA-Z0-9\_]{1,}$/)" 
         inner-label="Variable name"
         value="${config.name}"
-        hint-text="Enter the variable name that you would like displayed on all data outputs. Valid variable names start with a letter (a-z) with proceeding characters consisting of letters (a-z), underscore (_), dash (-), and numbers (0-9)."
+        hint-text="Enter the variable name that you would like displayed on all data outputs. Valid variable names start with a letter (a-z) with proceeding characters consisting of letters (a-z), underscore (_), and numbers (0-9)."
         required>
       </tangy-input>
       <tangy-input