+++ tangy-form-editor-add-input.js
@@ -56,43 +56,44 @@ class TangyFormEditorAddInput extends PolymerElement {
         <div>
           <h4>Inputs</h4>
           <div class="element-list">
-            <mwc-button icon="date_range" on-click="addThis"><span id="tangy-date-widget">Date</span></mwc-button><br>
-            <mwc-button icon="calendar_today" on-click="addThis"><span id="tangy-partial-date-widget">Partial Date</span></mwc-button><br>
-            <mwc-button icon="text_fields" on-click="addThis"><span id="tangy-text-widget">Text</span></mwc-button><br>
-            <mwc-button icon="timer" on-click="addThis"><span id="tangy-time-widget">Time</span></mwc-button><br>
-            <mwc-button icon="looks_one" on-click="addThis"><span id="tangy-number-widget">Number</span></mwc-button><br>
-            <mwc-button icon="email" on-click="addThis"><span id="tangy-email-widget">Email</span></mwc-button><br>
+            <mwc-button icon="date_range" on-click="addThis" id="tangy-date-widget">Date</mwc-button><br>
+            <mwc-button icon="calendar_today" on-click="addThis" id="tangy-partial-date-widget">Partial Date</mwc-button><br>
+            <mwc-button icon="text_fields" on-click="addThis" id="tangy-text-widget">Text</mwc-button><br>
+            <mwc-button icon="timer" on-click="addThis" id="tangy-time-widget">Time</mwc-button><br>
+            <mwc-button icon="looks_one" on-click="addThis" id="tangy-number-widget">Number</mwc-button><br>
+            <mwc-button icon="email" on-click="addThis" id="tangy-email-widget">Email</mwc-button><br>
           </div>
         </div>
         <div>
           <h4>Location</h4>
           <div class="element-list">
-              <mwc-button icon="add_location" on-click="addThis"><span id="tangy-gps-widget">GPS</span></mwc-button><br>
-              <mwc-button icon="location_city" on-click="addThis"><span id="tangy-location-widget">Location</span></mwc-button>
+              <mwc-button icon="add_location" on-click="addThis" id="tangy-gps-widget">GPS</mwc-button><br>
+              <mwc-button icon="location_city" on-click="addThis" id="tangy-location-widget">Location</mwc-button>
           </div>
         </div>
         <div>
           <h4>Lists</h4>
           <div class="element-list">
-              <mwc-button icon="check_box" on-click="addThis"><span id="tangy-checkbox-widget">Checkbox</span></mwc-button><br>
-              <mwc-button icon="check_box_outline_blank" on-click="addThis"><span id="tangy-checkboxes-widget">Checkbox Group</span></mwc-button><br>
-              <mwc-button icon="arrow_drop_down_circle" on-click="addThis"><span id="tangy-select-widget">Dropdown (select)</span></mwc-button><br>
-              <mwc-button icon="radio_button_unchecked" on-click="addThis"><span id="tangy-radio-buttons-widget">Radio Buttons</span></mwc-button><br>
+              <mwc-button icon="check_box" on-click="addThis" id="tangy-checkbox-widget">Checkbox</mwc-button><br>
+              <mwc-button icon="check_box_outline_blank" on-click="addThis" id="tangy-checkboxes-widget">Checkbox Group</mwc-button><br>
+              <mwc-button icon="arrow_drop_down_circle" on-click="addThis" id="tangy-select-widget">Dropdown (select)</mwc-button><br>
+              <mwc-button icon="radio_button_unchecked" on-click="addThis" id="tangy-radio-buttons-widget">Radio Buttons</mwc-button><br>
 
               </div>
         </div>
         <div>
           <h4>Misc</h4>
           <div class="element-list">
-            <mwc-button icon="image" on-click="addThis"><span id="tangy-image-widget">Image</span></mwc-button><br>
-            <mwc-button icon="code" on-click="addThis"><span id="tangy-box-widget">HTML content container</span></mwc-button><br>
-            <mwc-button icon="attach_money" on-click="addThis"><span id="tangy-template-widget">HTML template</span></mwc-button><br>
-            <mwc-button icon="filter_center_focus" on-click="addThis"><span id="tangy-qr-widget">QR Code Scanner</span></mwc-button><br>
-            <mwc-button icon="question_answer" on-click="addThis"><span id="tangy-eftouch-widget">EF Touch</span></mwc-button><br>
-            <mwc-button icon="speaker" on-click="addThis"><span id="tangy-acasi-widget">Acasi</span></mwc-button><br>
-            <mwc-button icon="av_timer"on-click="addThis"><span id="tangy-timed-widget" >Timed Grid</span></mwc-button><br>
-            <mwc-button icon="grid_on"on-click="addThis"><span id="tangy-untimed-grid-widget" >Untimed Grid</span></mwc-button><br>
-            <mwc-button icon="thumbs_up_down"on-click="addThis"><span id="tangy-consent-widget" >Consent</span></mwc-button><br>
+            <mwc-button icon="image" on-click="addThis" id="tangy-image-widget">Image</mwc-button><br>
+            <mwc-button icon="brush" on-click="addThis" id="tangy-signature-widget">Signature</mwc-button><br>
+            <mwc-button icon="code" on-click="addThis" id="tangy-box-widget">HTML content container</mwc-button><br>
+            <mwc-button icon="attach_money" on-click="addThis" id="tangy-template-widget">HTML template</mwc-button><br>
+            <mwc-button icon="filter_center_focus" on-click="addThis" id="tangy-qr-widget">QR Code Scanner</mwc-button><br>
+            <mwc-button icon="question_answer" on-click="addThis" id="tangy-eftouch-widget">EF Touch</mwc-button><br>
+            <mwc-button icon="speaker" on-click="addThis" id="tangy-acasi-widget">Acasi</mwc-button><br>
+            <mwc-button icon="av_timer"on-click="addThis" id="tangy-timed-widget" >Timed Grid</mwc-button><br>
+            <mwc-button icon="grid_on"on-click="addThis" id="tangy-untimed-grid-widget" >Untimed Grid</mwc-button><br>
+            <mwc-button icon="thumbs_up_down"on-click="addThis" id="tangy-consent-widget" >Consent</mwc-button><br>
           </div>
         </div>
       </div>